requirements.txt
openai>=1.0.0
elevenlabs>=0.2.0
python-dotenv>=1.0.0
moviepy>=1.0.3
ffmpeg-python>=0.2.0
requests>=2.31.0
pydub>=0.25.1
numpy>=1.24.0
Pillow>=10.0.0
sounddevice>=0.4.6
soundfile>=0.12.1
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
google-cloud-texttospeech>=2.14.0
langdetect>=1.0.9
translate>=3.6.1 